# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Laravel with PostgreSQL - Created Muhamad Ferdinal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pesanan-produk": {
            "get": {
                "tags": [
                    "Product Order"
                ],
                "summary": "Get list of product orders",
                "description": "Returns list of product orders",
                "operationId": "findAllProductOrder",
                "responses": {
                    "200": {
                        "description": "Pengambilan Data Pesanan Produk Berhasil!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data Pesanan Produk!"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product Order"
                ],
                "summary": "Create product order data",
                "description": "Create product order data",
                "operationId": "createProductOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pesanan_id",
                                    "produk_id",
                                    "jumlah"
                                ],
                                "properties": {
                                    "pesanan_id": {
                                        "description": "ID Pesanan (Required)",
                                        "type": "uuid"
                                    },
                                    "produk_id": {
                                        "description": "ID Produk (Required)",
                                        "type": "uuid"
                                    },
                                    "jumlah": {
                                        "description": "Jumlah Produk (Required)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Pesanan Produk Berhasil Dibuat!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Membuat Data Pesanan Produk!"
                    }
                }
            }
        },
        "/pesanan-produk/{pesanan_produk_id}": {
            "get": {
                "tags": [
                    "Product Order"
                ],
                "summary": "Get single product order data",
                "description": "Returns single product order data",
                "operationId": "findOneProductOrder",
                "parameters": [
                    {
                        "name": "pesanan_produk_id",
                        "in": "path",
                        "description": "Pesanan Produk id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengambilan Data Pesanan Produk Berhasil!"
                    },
                    "404": {
                        "description": "Data Pesanan Produk Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data Pesanan Produk!"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product Order"
                ],
                "summary": "Update product order data",
                "description": "Update product order data",
                "operationId": "updateOneProductOrder",
                "parameters": [
                    {
                        "name": "pesanan_produk_id",
                        "in": "path",
                        "description": "Pesanan Produk id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pesanan_id": {
                                        "description": "ID Pesanan (Optional)",
                                        "type": "uuid"
                                    },
                                    "produk_id": {
                                        "description": "ID Produk (Optional)",
                                        "type": "uuid"
                                    },
                                    "jumlah": {
                                        "description": "Jumlah Produk (Optional)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Pesanan Produk Berhasil Diperbarui!"
                    },
                    "404": {
                        "description": "Data Pesanan Produk Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Memperbarui Data Pesanan Produk!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product Order"
                ],
                "summary": "Delete product order data",
                "description": "Delete product order data",
                "operationId": "deleteOneProductOrder",
                "parameters": [
                    {
                        "name": "pesanan_produk_id",
                        "in": "path",
                        "description": "Pesanan Produk id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Pesanan Produk Berhasil Dihapus!"
                    },
                    "404": {
                        "description": "Data Pesanan Produk Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Menghapus Data Pesanan Produk!"
                    }
                }
            }
        },
...
    }
}